more better numbers, danger sensing spotlight

+++ OgreProject3/OgreProject3/LaneHandler.cpp
@@ -1,6 +1,6 @@
 #include "LaneHandler.h"
 
-void LaneHandler::init(SceneManager* sceneMgr, float laneLength, float leftBound, float rightBound, float speedMin, float speedMax, float spawnMin, float spawnMax, float pauseMin, float pauseMax)
+void LaneHandler::init(SceneManager* sceneMgr, float laneLength, float leftBound, float rightBound, float speedMin, float speedMax, float spawnMin, float spawnMax, int pauseMin, int pauseMax)
 {
 	this->laneLength = laneLength;
 	this->leftBound = leftBound;
